Remove old-school placard.
Apple has been advising against the use of placerds since the debut of the iOS style transparent scrollers. Instead of an old-school placrd, I've added a bar below the scroll view. Externally, the API should be the same. Internally, the view is now a subclass of vanilla.Group instead of vanilla.List.